adds auth0 id_token check
