+++ server.js
@@ -8,8 +8,27 @@ const TemplateTasksRouter = require("./data/templates-tasks/templates-tasks-rout
 const Weather = require("./data/weather/weather.router");
 // const Auth = require("./data/middleware/restriced.middleware")
 const delayLogsRouter = require("./data/delay-logs/delay_logs.router");
-const checkJwt = require('./auth/validation')
+const jwt = require("express-jwt");
+const jwksRsa = require("jwks-rsa");
 
+
+
+
+const checkJwt = jwt({
+  secret: jwksRsa.expressJwtSecret({
+    cache: true,
+    rateLimit: true,
+    jwksRequestsPerMinute: 5,
+    jwksUri: `https://gannondarcy2.auth0.com/.well-known/jwks.json`
+     
+  }),
+
+  // Validate the audience and the issuer.
+  audience: "OzMg1e7JDNF7DogxPEPvGzpG7fvvDHNe",
+  issuer: `https://gannondarcy2.auth0.com/`,
+  algorithms: ["RS256"],
+
+});
 
 
 
@@ -34,11 +53,11 @@ server.use(helmet());
 server.use(express.json());
 server.use(logger);
 // server.use('/auth',authRouter)
-server.use("/projects",checkJwt, ProjectsRouter);
-server.use("/projects/tasks",checkJwt,TasksRouter);
+server.use("/projects", ProjectsRouter,checkJwt);
+server.use("/projects/tasks",TasksRouter,checkJwt);
 server.use("/templates",TemplatesRouter);
 server.use("/projects/tasks/templates",TemplateTasksRouter);
 server.use("/weather", Weather);
-server.use("/delay_logs",checkJwt,delayLogsRouter);
+server.use("/delay_logs",delayLogsRouter,checkJwt);
 
 module.exports = server;
